[Kascade] Bugfix: Stream close before end with file output option.
Output of kascade is performed by sized queue and a thread which reads
the queue to write in file or process's input stream.  In file case
(FileOutput) the stream was directly closed by END instruction instead
of write 'nil' packet in the queue.  Therefore the last data written
this queue and does not written in file by the output thread when END
occured was lost.

This fix remove the close override of OutputFile and close correctly the
stream such as defined in mother class.

Change-Id: I9e41fcd860cc5e5745d725357ca851823e2e8213