[client] Bugfix : check if ENV['USER'] and ENV['USERNAME'] is set before call Etc.getlogin
getlogin calls getlogin function on unistd if it is defined. This
function returns the owner of tty. Therefore if an user would change the
effective id with su or sudo and launchs the client.  The client takes
the first user id. For example, if you connect as root with ssh and
use su to be frontend user, the client sends the root username instead
frontend to kadeploy.

The service ident catch the username mismatch and the kadeploy server returns 401 error.

Change-Id: I980e34d410c9c08abd87239f5ab2b6fb29dc89af