[Kaworkflow] bugfix: Add NodeSetId object which is a counter object to avoid inconsistent id after two splits of automata.
Before this patch: the counter of entry ":nodesets_id" in context was an integer. The context is cloned (only first level) when a task splits (0->1,2).
The problem occurs when one branch (1) splits again (1->3,4)the counter of another branch is not updated.
When the another branch (2) splits the id start with value of common split and reuse the id of the first branch split (2->3,4 instead of 2->5,6).

This patch replaces the interger by a counter (NodeSetId) which is common with all branchs.
When a branch updates it, the another branch can see this update and it can use a free identifier.

This patch fixes the first part of bug: https://intranet.grid5000.fr/bugzilla/show_bug.cgi?id=5462

Change-Id: I331cad97837371d9e93da0cb41db0543641e4a8a