[automata] Place the error message before the split message
This patch splits 'fail!' function in two parts: the error displaying and the actions on task object.
The error displaying comes inside a new function called: display_fail_message and it will start before split action.
The rest of 'fail!' function is called at end of step and after the split to make some stuff (depending of object).

Change-Id: Ifb03a20dfc8563687bc5bc54a5538072fe4a7c63