Dump to stderr instead of a file
Ruby delegates all I/Os to an 'I/O thread'. It might be possible that,
for some reason, that I/O thread is blocked (even if that should never
happen as Ruby is supposed to use non-blocking I/Os). As a result, the
dump file might not get written to disk.

To work around that, write to STDERR, and let the OS redirect the data
to a file.

This might fix the fact that kill -USR1 doesn't work when Kadeploy
encounters G5K bug 5484.

Change-Id: I12f94ca52840ffbbbda203c27b7a12101b6695e9