[Execute] Fix race condition in kill_recursive
The previous version of this code was racy: if a process
terminates during the execution of kill_recursive (between the
ps call to gather children, and the subsequent kill call),
then ESRCH is raised. This terminates the execution of kill_recursive,
since ESRCH is only rescued in the parent kill! method.

This patch fixes that by moving the rescueing of ESRCH to each kill
call.

Change-Id: I8f655bec41873f872fc726ffdca5ef193b6cbfa4