[Execute] Add possibility to launch a process without any pipe.
Before this patch, the execute function has a flag to disable stdin redirection.
This patch provides flag to disable stderr and stdout redirections.
This reduce the number of pipes.
With this patch all disabled redirections in new process are opened with /dev/null
file descriptor instead of kadeploy standard input/output.

Change-Id: I03f917d01fd185dfde8f18f50f8333bb4eebd245