[Execute] Add lock when execute reads or closes a file descriptor.
Execute checks if a file descriptor is closed before close or read it
without any lock. If another thread calls 'wait' function, it can close
a file descriptor when the first thread has just check the file
descriptor state.  So, the action of the first thread raises the
exception : 'IOError : already closed'.

Change-Id: Ic014d090ccda8237f681e8878c9e7dcd169874fd