[Httpd] Add servlet timeout to avoid the server lock.
The webrick server has pool of threads. Threrefore, when a servlet never
finishes, the used thread becomes busy.  When all threads are busy, the webrick
server can't respond to new request.  This timeout raise an exception after 10
min if the servlet has not finished. This exception shows where the thread was
locked.

Change-Id: Ic646ab22c33760e242238f3a3551eec74dd7e5ce