[Execute] bugfix: Find all opened FD for close after a fork and not only those opened by ruby.
The module Ruby/Mysql has a part in C language. This part manages the socket
which is connected to the server.  This socket is not referenced by ruby. When
we fork this socket is dumped to the child and we can not find this socket with
ObjectSpace.

So this patch replace ObjectSpace by "Dir.foreach('/proc/self/fd')". This line
iterate with all opened file descriptor for the current application.

Change-Id: I214638750dbdb6e7f3e1e6cd2768d1fcf959c072