[Execute] Refactoring of execute class to simplify the design and limit the race conditions.
With this patch, there is only one thread to manage the pipes and waitpid.
The wait of this function checks the status of this thread.
Killing a 'execute.wait()' function kills the process with 'execute.kill()' function.

This patch provides some test for execute class in test/test_execute.rb.

Change-Id: Id585094f462c86647b8b4ef3013e64b00badcd91