[Cache] Refactoring cache to simplify the design and limit the race conditions.
Before this patch the cache had useless functions and the object FileCache
could be re-index in the Cache data structure.  The size was not recomputed and
there was race condition on it.

This patch uses the monitor design pattern to simplify the design.  The cache
has a lock to add or remove elements. Every FileCache has a lock to update, acquire or
release a token, fetch a file etc.

This patch updates test/test_cache.rb to test this module.

Change-Id: I9a660ce179d090d4c449c40ca5a7b1a7b82707eb