[DB] Merge disconnect function into free function.
Before this patch, the end of database connection must be made in two times:
disconnect and free. The issue is that kaworkflow calls free without
disconnect.

Now free function closes the SQL connection if a connection was established
and disconnect function is deleted. This resolves the kaworkflow issue.

Note for developers:

Kadeploy establishes one SQL connexion on each action (in fact, each connexion
to server, it can be avoided for the polling if workflow keeps the user right
only for the polling).  When a workflow is created, a second SQL connection is
established because the workflow has longer life than execution context.

My opinion is to made only one connection for everything. I was tested with the
modification of file server.rb: check_database establishes the SQL connection
and database_handler return the connection. It works fine with concurrency.
But EJ said that the MySQL module has weird behavior after one week of running.

Change-Id: I2b1610890aded0b538fade4a3fc386857b777127