[kadeploy3d] bugfix: display 'nil' if a backtrace of thread is nil.
When a USR2 signal was received, Kadeploy explores all thread and display a backtrace.
A backtrace of a thread could be nil. Before this patch, we try to call join even if this backtrace is nil.

This patch resolve this issue.

See bug https://intranet.grid5000.fr/bugzilla/show_bug.cgi?id=5650 for
details.

Change-Id: Iac7831450a983aaa3e52e71266285e87f44ce057