[Cache] Change ordinal references by set of wid.
Before this patch, if kadeploy kills a workflow during it grabs the file, there
is non-null probability that a file is cached but it not stored to the workflow
structure. In this case, Kadeploy has lost the token and it can't return it to the cache.

This patch replace the ordinal reference by workflow id. When kadeploy kills a
workflow, all token take by wid is returned.

Change-Id: I8df4b98976ff4b0a6837ff1c08feb4906394c67b