[Kaenv] Change API to be RFC3986 compliant.
Before this patch, the optional elements were in the path. So, /environments/toto and /environments//toto are different.
However RFC3986 does not allow empty directories inside a URI.

This patch transforms the API to use the URI query instead of URI path, with the following keywords username, user and version.
For instance the both examples above become /environments?username=toto and /environments?name=toto

Change-Id: Icc0b3cd777f5f769da4eedd881a9e600ac10447f